review fix: structure, proptypes, counter
