+++ src/components/burger-constructor/burger-constructor.jsx
@@ -1,155 +1,16 @@
-import React, { useState, useRef, useCallback } from 'react';
+import React, { useCallback } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
-import { useDrop, useDrag } from 'react-dnd';
-import {
-  CurrencyIcon,
-  Button,
-  ConstructorElement,
-  DragIcon
-} from '@ya.praktikum/react-developer-burger-ui-components';
+import { useDrop } from 'react-dnd';
+import { ConstructorElement } from '@ya.praktikum/react-developer-burger-ui-components';
 import styleConstructor from './burger-constructor.module.scss';
 import { Scrollbars } from 'react-custom-scrollbars';
 import clsx from 'clsx';
-import PropTypes from 'prop-types';
 import { v4 as uuidv4 } from 'uuid';
 
-import {
-  addIngredient,
-  DELETE_INGREDIENT,
-  MOVE_INGREDIENT
-} from '../../services/actions/constructor-actions';
+import { addIngredient, MOVE_INGREDIENT } from '../../services/actions/constructor-actions';
 
-import OrderDetails from '../order-details/order-details';
-import Modal from '../modal/modal';
-import { getOrder } from '../../services/actions/order-actions';
-import dataIngredientsType from '../../utils/types';
-
-const Total = () => {
-  const totalPrice = useSelector(
-    (store) =>
-      (store.burgerConstructor.constructorBun
-        ? store.burgerConstructor.constructorBun.price * 2
-        : 0) +
-      (store.burgerConstructor.constructorItems
-        ? store.burgerConstructor.constructorItems.reduce((total, item) => total + item.price, 0)
-        : 0)
-  );
-
-  const { constructorBun, constructorItems } = useSelector((store) => ({
-    constructorBun: store.burgerConstructor.constructorBun,
-    constructorItems: store.burgerConstructor.constructorItems
-  }));
-
-  const [modalVisibility, setModalVisibility] = useState(false);
-
-  const dispatch = useDispatch();
-
-  const makeOrder = () => {
-    if (constructorBun) {
-      setModalVisibility(true);
-      const ingredients = [constructorBun._id, ...constructorItems.map((item) => item._id)];
-      dispatch(getOrder(ingredients));
-    }
-  };
-
-  const handleCloseModal = () => {
-    setModalVisibility(false);
-  };
-
-  return (
-    <div className={clsx(styleConstructor.total, 'pt-6 pr-4')}>
-      <div className={clsx(styleConstructor.total_price, 'mr-10')}>
-        <span className='text text_type_digits-medium mr-2'>{totalPrice}</span>
-        <CurrencyIcon type='primary' />
-      </div>
-      <Button type='primary' size='large' onClick={makeOrder}>
-        Оформить заказ
-      </Button>
-      {modalVisibility && (
-        <Modal closeModal={handleCloseModal}>
-          <OrderDetails />
-        </Modal>
-      )}
-    </div>
-  );
-};
-
-const ConstructorItem = ({ ingredient, index, moveItem }) => {
-  const { position } = ingredient;
-
-  const ref = useRef(null);
-  const [{ handlerId }, drop] = useDrop({
-    accept: 'constructorItem',
-    collect (monitor) {
-      return {
-        handlerId: monitor.getHandlerId()
-      };
-    },
-    hover (item, monitor) {
-      if (!ref.current) {
-        return;
-      }
-      const dragIndex = item.index;
-      const hoverIndex = index;
-
-      if (dragIndex === hoverIndex) {
-        return;
-      }
-      const hoverBoundingRect = ref.current?.getBoundingClientRect();
-      const hoverMiddleY = (hoverBoundingRect.bottom - hoverBoundingRect.top) / 2;
-      const clientOffset = monitor.getClientOffset();
-      const hoverClientY = clientOffset.y - hoverBoundingRect.top;
-
-      if (dragIndex < hoverIndex && hoverClientY < hoverMiddleY) {
-        return;
-      }
-      if (dragIndex > hoverIndex && hoverClientY > hoverMiddleY) {
-        return;
-      }
-      moveItem(dragIndex, hoverIndex);
-      item.index = hoverIndex;
-    }
-  });
-
-  const [{ isDragging }, drag] = useDrag({
-    type: 'constructorItem',
-    item: () => {
-      return { position, index };
-    },
-    collect: (monitor) => ({
-      isDragging: monitor.isDragging()
-    })
-  });
-
-  drag(drop(ref));
-
-  const deleteIngredient = (position) => {
-    dispatch({
-      type: DELETE_INGREDIENT,
-      position
-    });
-  };
-
-  const dispatch = useDispatch();
-
-  return (
-    <div
-      className={clsx(styleConstructor.drag_element, isDragging ? styleConstructor.hide : '')}
-      ref={ref}
-      data-handler-id={handlerId}
-    >
-      <div className={styleConstructor.drag_icon}>
-        <DragIcon type='primary' />
-      </div>
-      <ConstructorElement
-        text={ingredient.name}
-        price={ingredient.price}
-        thumbnail={ingredient.image}
-        handleClose={() => deleteIngredient(ingredient.position)}
-      />
-    </div>
-  );
-};
+import Total from './total/total';
+import ConstructorItem from './constructor-item/constructor-item';
 
 const BurgerConstructor = () => {
   const { constructorBun, constructorItems } = useSelector((store) => ({
@@ -264,12 +125,6 @@ const BurgerConstructor = () => {
       <Total />
     </section>
   );
-};
-
-ConstructorItem.propTypes = {
-  ingredient: dataIngredientsType.isRequired,
-  index: PropTypes.number.isRequired,
-  moveItem: PropTypes.func.isRequired
 };
 
 export default BurgerConstructor;
