+++ src/components/burger-ingredients/burger-ingredients.module.scss
@@ -16,50 +16,6 @@
   height: 40px;
 }
 
-.tabs {
-  display: flex;
-}
-
-.catalog {
-  display: flex;
-  justify-content: space-between;
-  flex-wrap: wrap;
-  row-gap: 16px;
-}
-
-.card {
-  position: relative;
-  width: 272px;
-  display: flex;
-  justify-content: center;
-  flex-direction: column;
-  background-color: $bg;
-  border-radius: 20px;
-}
-
-
-.card:hover {
-  cursor: grab;
-  outline: 2px solid $input;
-}
-
-.card_price {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.card_name {
-  text-align: center;
-  height: 48px;
-  text-overflow: ellipsis;
-}
-
-.counter {
-  position: absolute;
-  top: 0;
-  right: 0;
-}
 
 .track_vertical {
   position: absolute;
