+++ src/components/burger-ingredients/ingredient-card/ingredient-card.module.scss
@@ -1,0 +1,34 @@
+@import '../../../css/variables.scss';
+
+.card {
+  position: relative;
+  width: 272px;
+  display: flex;
+  justify-content: center;
+  flex-direction: column;
+  background-color: $bg;
+  border-radius: 20px;
+}
+
+.card:hover {
+  cursor: grab;
+  outline: 2px solid $input;
+}
+
+.card_price {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+.card_name {
+  text-align: center;
+  height: 48px;
+  text-overflow: ellipsis;
+}
+
+.counter {
+  position: absolute;
+  top: 0;
+  right: 0;
+}
