+++ src/components/burger-ingredients/tabs/tabs.module.scss
@@ -1,0 +1,5 @@
+@import '../../../css/variables.scss';
+
+.tabs {
+  display: flex;
+}
